Split models, and put classes depending on its purpose
