+++ src/app/models/proceedings/ProceedingsDate.java
@@ -1,6 +1,8 @@
 package app.models.proceedings;
 
 import java.time.LocalDate;
+
+import app.models.ValueObject;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -8,6 +10,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
+@ValueObject
 public class ProceedingsDate {
 
     private LocalDate creationDate;
