+++ src/app/models/records/DocumentType.java
@@ -1,0 +1,13 @@
+package app.models.records;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public class DocumentType {
+    private String name;
+    private int code;
+}
