+++ src/app/models/records/parts/NaturalPerson.java
@@ -1,4 +1,4 @@
-package app.models;
+package app.models.records.parts;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
