+++ src/app/models/records/parts/StateEntity.java
@@ -1,4 +1,4 @@
-package app.models;
+package app.models.records.parts;
 
 public class StateEntity extends Person {
