+++ src/app/models/users/User.java
@@ -1,0 +1,14 @@
+package app.models.users;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public abstract class User {
+
+    private String name;
+    private String password;
+}
