+++ src/app/services/UserService.java
@@ -2,9 +2,9 @@ package app.services;
 
 import app.database.ConnectionManager;
 import app.database.UserRepository;
-import app.models.Coordinator;
-import app.models.SuperUser;
-import app.models.User;
+import app.models.users.Coordinator;
+import app.models.users.SuperUser;
+import app.models.users.User;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
