+++ src/test/controllers/UserControllerTest.java
@@ -1,9 +1,9 @@
 package test.controllers;
 
 import app.controllers.UserController;
-import app.models.Coordinator;
-import app.models.SuperUser;
-import app.models.User;
+import app.models.users.Coordinator;
+import app.models.users.SuperUser;
+import app.models.users.User;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
